Update Pytest configuration and slow test handling
Cleanups of unused utilities imports

slow tests moved to a different pytest way of marking
