+++ test/python/pytest.ini
@@ -1,2 +1,5 @@
 [pytest]
+markers =
+    slow_test: mark test as slow
+
 pythonpath = ../../build/python
