+++ test/python/test_mixture.py
@@ -1,6 +1,5 @@
 import pytest
 import cantera as ct
-from . import utilities
 from .utilities import (
     assertNear,
     assertArrayNear
