+++ test/python/test_purefluid.py
@@ -3,7 +3,6 @@ import numpy as np
 import pytest
 
 import cantera as ct
-from . import utilities
 from .utilities import (
     assertNear,
 )
